+++ src/gui/menu_item.cpp
@@ -56,8 +56,9 @@ void
 MenuItem::draw(DrawingContext& context, const Vector& pos, int menu_width, bool active)
 {
   context.color().draw_text(Resources::normal_font, text,
-                              Vector( pos.x + menu_width/2 , pos.y - int(Resources::normal_font->get_height())/2 ),
-                              ALIGN_CENTER, LAYER_GUI, active ? ColorScheme::Menu::active_color : get_color());
+                            Vector( pos.x + static_cast<float>(menu_width) / 2.0f,
+                                    pos.y - static_cast<float>(Resources::normal_font->get_height()) / 2.0f ),
+                            ALIGN_CENTER, LAYER_GUI, active ? ColorScheme::Menu::active_color : get_color());
 }
 
 Color
