+++ src/object/background.cpp
@@ -248,10 +248,10 @@ Background::draw_image(DrawingContext& context, const Vector& pos_)
   Sizef parallax_image_size = (1.0f - speed) * screen + level * speed;
   Rectf cliprect = context.get_cliprect();
 
-  int start_x = static_cast<int>(floorf((cliprect.get_left()  - (pos_.x - image->get_width() /2.0f)) / static_cast<float>(image->get_width())));
-  int end_x   = static_cast<int>(ceilf((cliprect.get_right()  - (pos_.x + image->get_width() /2.0f)) / static_cast<float>(image->get_width()))) + 1;
-  int start_y = static_cast<int>(floorf((cliprect.get_top()   - (pos_.y - image->get_height()/2.0f)) / static_cast<float>(image->get_height())));
-  int end_y   = static_cast<int>(ceilf((cliprect.get_bottom() - (pos_.y + image->get_height()/2.0f)) / static_cast<float>(image->get_height()))) + 1;
+  int start_x = static_cast<int>(floorf((cliprect.get_left()  - (pos_.x - static_cast<float>(image->get_width()) /2.0f)) / static_cast<float>(image->get_width())));
+  int end_x   = static_cast<int>(ceilf((cliprect.get_right()  - (pos_.x + static_cast<float>(image->get_width()) /2.0f)) / static_cast<float>(image->get_width()))) + 1;
+  int start_y = static_cast<int>(floorf((cliprect.get_top()   - (pos_.y - static_cast<float>(image->get_height())/2.0f)) / static_cast<float>(image->get_height())));
+  int end_y   = static_cast<int>(ceilf((cliprect.get_bottom() - (pos_.y + static_cast<float>(image->get_height())/2.0f)) / static_cast<float>(image->get_height()))) + 1;
 
   switch(alignment)
   {
@@ -259,7 +259,7 @@ Background::draw_image(DrawingContext& context, const Vector& pos_)
       for(int y = start_y; y < end_y; ++y)
       {
         Vector p(pos_.x - parallax_image_size.width / 2.0f,
-                 pos_.y + y * image->get_height()  - image->get_height() / 2.0f);
+                 pos_.y + static_cast<float>(y) * static_cast<float>(image->get_height()) - static_cast<float>(image->get_height()) / 2.0f);
         context.color().draw_surface(image, p, layer);
       }
       break;
@@ -267,8 +267,8 @@ Background::draw_image(DrawingContext& context, const Vector& pos_)
     case RIGHT_ALIGNMENT:
       for(int y = start_y; y < end_y; ++y)
       {
-        Vector p(pos_.x + parallax_image_size.width / 2.0f - image->get_width(),
-                 pos_.y + y * image->get_height() - image->get_height() / 2.0f);
+        Vector p(pos_.x + parallax_image_size.width / 2.0f - static_cast<float>(image->get_width()),
+                 pos_.y + static_cast<float>(y) * static_cast<float>(image->get_height()) - static_cast<float>(image->get_height()) / 2.0f);
         context.color().draw_surface(image, p, layer);
       }
       break;
@@ -276,7 +276,7 @@ Background::draw_image(DrawingContext& context, const Vector& pos_)
     case TOP_ALIGNMENT:
       for(int x = start_x; x < end_x; ++x)
       {
-        Vector p(pos_.x + x * image->get_width() - image->get_width() / 2.0f,
+        Vector p(pos_.x + static_cast<float>(x) * static_cast<float>(image->get_width()) - static_cast<float>(image->get_width()) / 2.0f,
                  pos_.y - parallax_image_size.height / 2.0f);
         context.color().draw_surface(image, p, layer);
       }
@@ -285,8 +285,8 @@ Background::draw_image(DrawingContext& context, const Vector& pos_)
     case BOTTOM_ALIGNMENT:
       for(int x = start_x; x < end_x; ++x)
       {
-        Vector p(pos_.x + x * image->get_width()  - image->get_width() / 2.0f,
-                 pos_.y - image->get_height() + parallax_image_size.height / 2.0f);
+        Vector p(pos_.x + static_cast<float>(x) * static_cast<float>(image->get_width()) - static_cast<float>(image->get_width()) / 2.0f,
+                 pos_.y - static_cast<float>(image->get_height()) + parallax_image_size.height / 2.0f);
         context.color().draw_surface(image, p, layer);
       }
       break;
@@ -295,8 +295,8 @@ Background::draw_image(DrawingContext& context, const Vector& pos_)
       for(int y = start_y; y < end_y; ++y)
         for(int x = start_x; x < end_x; ++x)
         {
-          Vector p(pos_.x + x * image->get_width()  - image->get_width()/2,
-                   pos_.y + y * image->get_height() - image->get_height()/2);
+          Vector p(pos_.x + static_cast<float>(x) * static_cast<float>(image->get_width()) - static_cast<float>(image->get_width()) / 2.0f,
+                   pos_.y + static_cast<float>(y) * static_cast<float>(image->get_height()) - static_cast<float>(image->get_height()) / 2.0f);
 
           if (image_top.get() != NULL && (y < 0))
           {
@@ -326,7 +326,8 @@ Background::draw(DrawingContext& context)
 
   Sizef level_size(Sector::current()->get_width(),
                    Sector::current()->get_height());
-  Sizef screen(context.get_width(), context.get_height());
+  Sizef screen(static_cast<float>(context.get_width()),
+               static_cast<float>(context.get_height()));
   Sizef translation_range = level_size - screen;
   Vector center_offset(context.get_translation().x - translation_range.width  / 2.0f,
                        context.get_translation().y - translation_range.height / 2.0f);
